refactor: add separate folders for services, models and components
