+++ src/main/webapp/app/tasks/components/task-list/task-list.component.scss
