+++ src/main/webapp/app/tasks/components/task-list/task-list.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { of } from 'rxjs';
 
-import { TaskService } from '../task.service';
+import { TaskService } from '../../task.service';
 import { TaskListComponent } from './task-list.component';
 
 describe('TaskListComponent', () => {
