+++ src/main/webapp/app/tasks/models/task-file.model.ts
@@ -1,0 +1,7 @@
+export interface TaskFile {
+    id: string;
+    taskId: string;
+    fileName: string;
+    fileType: string;
+    fileSize: number;
+  }
