+++ src/main/webapp/app/tasks/models/task-request.model.ts
@@ -1,0 +1,7 @@
+/**
+ * A tiny task request.
+ */
+export interface TaskRequest {
+  name: string;
+  dueDate?: string;
+}
