+++ src/main/webapp/app/tasks/models/task.model.ts
@@ -1,0 +1,11 @@
+import { TaskFile } from "./task-file.model";
+
+/**
+ * A tiny task.
+ */
+export interface Task {
+  id: string;
+  name: string;
+  dueDate?: string;
+  files?: TaskFile[];
+}
